+++ .ci/build_dev_indexed_gzip_fileobj_fork_epicfaace.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 #
-# Build a test version of indexed_gzip.
+# Build a test version of indexed_gzip_fileobj_fork_epicfaace.
 
 set -e
 
@@ -11,6 +11,6 @@ source $thisdir/activate_env.sh "$envdir"
 
 # enable line tracing for cython
 # modules - see setup.py
-export INDEXED_GZIP_TESTING=1
+export indexed_gzip_fileobj_fork_epicfaace_TESTING=1
 
 python setup.py develop