+++ .ci/run_tests.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 #
-# Run indexed_gzip unit tests. Assumes that
+# Run indexed_gzip_fileobj_fork_epicfaace unit tests. Assumes that
 # python setup.py develop has been run.
 
 set -e
@@ -23,19 +23,8 @@ if [[ -n "$NELEMS" ]]; then
   NELEMS="--nelems $NELEMS"
 fi
 
-# No coverage on windows, because coverage or
-# pytest-cov seem to have trouble with threading/
-# multiproc, which causes the coverage report
-# generation to sporadically fail
-#
-# https://github.com/pytest-dev/pytest-cov/issues/406
-if [[ "$PLATFORM" == "windows"* ]]; then
-  EXTRA_ARGS="$EXTRA_ARGS --no-cov"
-fi
-
-python -m indexed_gzip.tests      \
-       -c setup.cfg               \
-       --cov-config=./.coveragerc \
+python -m indexed_gzip_fileobj_fork_epicfaace.tests      \
+       -c pyproject.toml          \
        -v -s                      \
        -m "$TEST_SUITE"           \
        -k "$TEST_PATTERN"         \