+++ indexed_gzip_fileobj_fork_epicfaace/tests/ctest_zran.pyx
@@ -23,6 +23,14 @@ import                    contextlib
 
 import numpy as np
 
+# The stdlib resource module is only
+# available on unix-like platforms.
+try:
+    import resource
+except ImportError:
+    resource = None
+
+
 cimport numpy as np
 
 from posix.types cimport  off_t
@@ -34,6 +42,8 @@ from libc.stdio  cimport (SEEK_SET,
                           FILE,
                           fdopen,
                           fwrite)
+
+from libc.stdint cimport int64_t
 
 from libc.string cimport memset, memcmp
 
@@ -52,15 +62,15 @@ from posix.mman cimport (mmap,
                          MAP_SHARED)
 
 
-from . import poll, check_data_valid, tempdir
+from . import poll, check_data_valid, tempdir, compress_inmem
 
 
 cdef extern from "sys/mman.h":
     cdef enum:
         MAP_FAILED
 
-cimport indexed_gzip.zran as zran
-cimport indexed_gzip.zran_file_util as zran_file_util
+cimport indexed_gzip_fileobj_fork_epicfaace.zran as zran
+cimport indexed_gzip_fileobj_fork_epicfaace.zran_file_util as zran_file_util
 
 np.import_array()
 
@@ -357,6 +367,14 @@ def test_getc():
         assert zran_file_util._getc_python(<PyObject*>f) == -1
         assert zran_file_util._ferror_python(<PyObject*>f) == 1
         PyErr_Clear()
+
+
+def test_seekable():
+
+    f = BytesIO(b"abc")
+    assert zran_file_util._seekable_python(<PyObject*>f) == 1
+    f.seekable = lambda: False
+    assert zran_file_util._seekable_python(<PyObject*>f) == 0
 
 
 def test_init(testfile, no_fds):
@@ -799,13 +817,14 @@ def test_seek_then_read_block(testfile, no_fds, nelems, niters, seed, use_mmap):
     with open(testfile, 'rb') as pyfid:
         cfid = fdopen(pyfid.fileno(), 'rb')
 
-        assert not zran.zran_init(&index,
-                                  NULL if no_fds else cfid,
-                                  <PyObject*>pyfid if no_fds else NULL,
-                                  indexSpacing,
-                                  32768,
-                                  131072,
-                                  zran.ZRAN_AUTO_BUILD)
+        ret = zran.zran_init(&index,
+                             NULL if no_fds else cfid,
+                             <PyObject*>pyfid if no_fds else NULL,
+                             indexSpacing,
+                             32768,
+                             131072,
+                             zran.ZRAN_AUTO_BUILD)
+        assert not ret, ret
 
         for i, se in enumerate(seekelems):
 
@@ -1018,12 +1037,43 @@ def test_readbuf_spacing_sizes(testfile, no_fds, nelems, niters, seed):
             print()
             zran.zran_free(&index)
 
-def test_export_then_import(testfile, no_fds):
 
-    cdef zran.zran_index_t  index1
-    cdef zran.zran_index_t  index2
+cdef _compare_indexes(zran.zran_index_t *index1,
+                      zran.zran_index_t *index2):
+    """Check that two indexes are equivalent. """
     cdef zran.zran_point_t *p1
     cdef zran.zran_point_t *p2
+
+    assert index2.compressed_size   == index1.compressed_size
+    assert index2.uncompressed_size == index1.uncompressed_size
+    assert index2.spacing           == index1.spacing
+    assert index2.window_size       == index1.window_size
+    assert index2.npoints           == index1.npoints
+
+    ws = index1.window_size
+
+    for i in range(index1.npoints):
+
+        p1 = &index1.list[i]
+        p2 = &index2.list[i]
+        msg = 'Error at point %d' % i
+
+        assert p2.cmp_offset   == p1.cmp_offset, msg
+        assert p2.uncmp_offset == p1.uncmp_offset, msg
+        assert p2.bits         == p1.bits, msg
+        if (not p1.data):
+            assert p1.data == p2.data, msg
+        else:
+            assert not memcmp(p2.data, p1.data, ws), msg
+
+
+def test_export_then_import(testfile, no_fds):
+    """Export-import round trip . Test exporting an index, then importing it
+    back in.
+    """
+
+    cdef zran.zran_index_t index1
+    cdef zran.zran_index_t index2
 
     indexSpacing = 1048576
     windowSize   = 32768
@@ -1063,36 +1113,16 @@ def test_export_then_import(testfile, no_fds):
             ret  = zran.zran_import_index(&index2, NULL if no_fds else cfid, <PyObject*>pyexportfid if no_fds else NULL)
             assert not ret, str(ret)
 
-        assert index2.compressed_size   == index1.compressed_size
-        assert index2.uncompressed_size == index1.uncompressed_size
-        assert index2.spacing           == index1.spacing
-        assert index2.window_size       == index1.window_size
-        assert index2.npoints           == index1.npoints
-
-        ws = index1.window_size
-
-        for i in range(index1.npoints):
-
-            p1 = &index1.list[i]
-            p2 = &index2.list[i]
-            msg = 'Error at point %d' % i
-
-            assert p2.cmp_offset   == p1.cmp_offset, msg
-            assert p2.uncmp_offset == p1.uncmp_offset, msg
-            assert p2.bits         == p1.bits, msg
-            if i == 0:
-                assert not p2.data, msg
-                assert not p1.data, msg
-            else:
-                assert p2.data, msg
-                assert p1.data, msg
-                assert not memcmp(p2.data, p1.data, ws), msg
+        _compare_indexes(&index1, &index2)
 
         zran.zran_free(&index1)
         zran.zran_free(&index2)
 
 
 def test_export_import_no_points(no_fds):
+    """Test exporting and importing an index which does not contain any
+    seek points.
+    """
     cdef zran.zran_index_t index
     cdef void             *buffer
 
@@ -1144,3 +1174,331 @@ def test_export_import_no_points(no_fds):
             pybuf = <bytes>(<char *>buffer)[:100]
             assert np.all(np.frombuffer(pybuf, dtype=np.uint8) == data)
             zran.zran_free(&index)
+
+
+def test_export_import_format_v0():
+    """Test index export and import on a version 0 index file. """
+
+    cdef zran.zran_index_t index1
+    cdef zran.zran_index_t index2
+    cdef int               ret
+
+    data = np.random.randint(1, 255, 1000000, dtype=np.uint8)
+    with tempdir():
+
+        with gzip.open('data.gz', 'wb') as f:
+            f.write(data.tostring())
+
+        with open('data.gz', 'rb')  as pyfid:
+            cfid = fdopen(pyfid.fileno(), 'rb')
+            assert not zran.zran_init(
+                &index1, cfid, NULL, 50000, 32768, 131072, 0)
+
+            assert not zran.zran_build_index(&index1, 0, 0)
+            _write_index_file_v0(&index1, 'data.gz.index')
+
+        with open('data.gz', 'rb')  as pyfid:
+            cfid = fdopen(pyfid.fileno(), 'rb')
+            assert not zran.zran_init(
+                &index2, cfid, NULL, 50000, 32768, 131072, 0)
+
+            with open('data.gz.index', 'rb') as pyidxfid:
+                cidxfid = fdopen(pyidxfid.fileno(), 'rb')
+                ret = zran.zran_import_index(&index2, cidxfid,  NULL)
+                assert ret == 0, ret
+
+        _compare_indexes(&index1, &index2)
+        zran.zran_free(&index1)
+        zran.zran_free(&index2)
+
+
+cdef _write_index_file_v0(zran.zran_index_t *index, dest):
+    """Write the given index out to a file, index file version 0 format. """
+
+    cdef zran.zran_point_t *point
+
+    with open(dest, 'wb') as f:
+        f.write(b'GZIDX\0\0')
+        f.write(<bytes>(<char *>(&index.compressed_size))[:8])
+        f.write(<bytes>(<char *>(&index.uncompressed_size))[:8])
+        f.write(<bytes>(<char *>(&index.spacing))[:4])
+        f.write(<bytes>(<char *>(&index.window_size))[:4])
+        f.write(<bytes>(<char *>(&index.npoints))[:4])
+
+        for i in range(index.npoints):
+            point = &index.list[i]
+            f.write(<bytes>(<char *>(&point.cmp_offset))[:8])
+            f.write(<bytes>(<char *>(&point.uncmp_offset))[:8])
+            f.write(<bytes>(<char *>(&point.bits))[:1])
+
+        for i in range(1, index.npoints):
+            point = &index.list[i]
+            data  = <bytes>point.data[:index.window_size]
+            f.write(data)
+
+
+def test_crc_validation(concat):
+    """Basic test of CRC validation. """
+
+    cdef zran.zran_index_t index
+    cdef void             *buffer
+    cdef int64_t           ret
+
+    # use uint32 so there are lots of zeros,
+    # and so there is something to compress
+    dsize             = 1048576 * 10
+    data              = np.random.randint(0, 255, dsize // 4, dtype=np.uint32)
+    cmpdata, strmoffs = compress_inmem(data.tobytes(), concat)
+    buf               = ReadBuffer(dsize)
+    buffer            = buf.buffer
+    f                 = [None]  # to prevent gc
+
+    def _zran_init(flags):
+        f[0] = BytesIO(cmpdata)
+        assert not zran.zran_init(&index,
+                                  NULL,
+                                  <PyObject*>f[0],
+                                  1048576,
+                                  32768,
+                                  131072,
+                                  flags)
+
+    def _run_crc_tests(shouldpass, flags=zran.ZRAN_AUTO_BUILD):
+        if shouldpass:
+            expect_build = zran.ZRAN_BUILD_INDEX_OK
+            expect_seek  = zran.ZRAN_SEEK_OK
+            expect_read  = dsize
+        else:
+            expect_build = zran.ZRAN_BUILD_INDEX_CRC_ERROR
+            expect_seek  = zran.ZRAN_SEEK_CRC_ERROR
+            expect_read  = zran.ZRAN_READ_CRC_ERROR
+
+        # CRC validation should occur on the first
+        # pass through a gzip stream, regardless
+        # of how that pass is initiated. Below we
+        # test the most common scenarios.
+
+        # Error if we try to build an index.  Note
+        # that an error here is not guaranteed, as
+        # the _zran_expand_index might need a few
+        # passes through the data to reach the end,
+        # which might cause inflation to be
+        # re-initialised, and therefore validation
+        # to be disabled.  It depends on the data,
+        # and on the constants used in
+        # _zran_estimate_offset
+        _zran_init(flags)
+        ret = zran.zran_build_index(&index, 0, 0)
+        assert ret == expect_build, ret
+        zran.zran_free(&index)
+
+        # error if we try to seek
+        _zran_init(flags)
+        ret = zran.zran_seek(&index, dsize - 1, SEEK_SET, NULL)
+        assert ret == expect_seek, ret
+        zran.zran_free(&index)
+
+        # error if we try to read
+        _zran_init(flags)
+        ret = zran.zran_read(&index, buffer, dsize)
+        assert ret == expect_read, ret
+        zran.zran_free(&index)
+
+        if shouldpass:
+            pybuf = <bytes>(<char *>buffer)[:dsize]
+            assert np.all(np.frombuffer(pybuf, dtype=np.uint32) == data)
+
+    def wrap(val):
+        return val % 255
+
+    # data/crc is good, all should be well
+    _run_crc_tests(True)
+
+    # corrupt the size, we should get an error
+    cmpdata[-1] = wrap(cmpdata[-1] + 1)  # corrupt size
+    _run_crc_tests(False)
+
+    # corrupt the crc, we should get an error
+    cmpdata[-1] = wrap(cmpdata[-1] - 1)  # restore size to correct value
+    cmpdata[-5] = wrap(cmpdata[-5] + 1)  # corrupt crc
+    _run_crc_tests(False)
+
+    # Corrupt a different stream, if we have more than one
+    cmpdata[-5] = wrap(cmpdata[-5] - 1)  # restore crc to correct value
+    if len(strmoffs) > 1:
+        for off in strmoffs[1:]:
+            cmpdata[off-1] = wrap(cmpdata[off-1] + 1)
+            _run_crc_tests(False)
+            cmpdata[off-1] = wrap(cmpdata[off-1] - 1)
+
+    # Disable CRC, all should be well, even with a corrupt CRC/size
+    # First test with good data
+    _run_crc_tests(True, zran.ZRAN_AUTO_BUILD | zran.ZRAN_SKIP_CRC_CHECK)
+
+    cmpdata[-1] = wrap(cmpdata[-1] + 1)  # corrupt size
+    _run_crc_tests(True, zran.ZRAN_AUTO_BUILD | zran.ZRAN_SKIP_CRC_CHECK)
+
+    cmpdata[-1] = wrap(cmpdata[-1] - 1)  # restore size to correct value
+    cmpdata[-5] = wrap(cmpdata[-5] - 1)  # corrupt crc
+    _run_crc_tests(True, zran.ZRAN_AUTO_BUILD | zran.ZRAN_SKIP_CRC_CHECK)
+
+
+def test_standard_usage_with_null_padding(concat):
+    """Make sure standard usage works with files that have null-padding after
+    the GZIP footer.
+
+    See https://www.gnu.org/software/gzip/manual/gzip.html#Tapes
+    """
+    cdef zran.zran_index_t index
+    cdef void             *buffer
+    cdef int64_t           ret
+
+    # use uint32 so there are lots of zeros,
+    # and so there is something to compress
+    dsize             = 1048576 * 10
+    data              = np.random.randint(0, 255, dsize // 4, dtype=np.uint32)
+    cmpdata, strmoffs = compress_inmem(data.tobytes(), concat)
+    buf               = ReadBuffer(dsize)
+    buffer            = buf.buffer
+    f                 = [None]  # to prevent gc
+
+    # random amount of padding for each stream
+    padding = np.random.randint(1, 100, len(strmoffs))
+
+    # new compressed data - bytearrays
+    # are initialised to contain all 0s
+    paddedcmpdata = bytearray(len(cmpdata) + padding.sum())
+
+    # copy old unpadded compressed data
+    # into new padded compressed data
+    padoff = 0  # offset into padded data
+    last   = 0  # offset to end of last copied stream in unpadded data
+    print('Padding streams [orig size: {}] ...'.format(len(cmpdata)))
+    for off, pad in zip(strmoffs, padding):
+
+        strm = cmpdata[last:off]
+
+        paddedcmpdata[padoff:padoff + len(strm)] = strm
+
+        print('  Copied stream from [{} - {}] to [{} - {}] ({} '
+              'padding bytes)'.format(
+                  last, off, padoff, padoff + len(strm), pad))
+
+        padoff += len(strm) + pad
+        last    = off
+
+
+    def _zran_init():
+        f[0] = BytesIO(paddedcmpdata)
+        assert not zran.zran_init(&index,
+                                  NULL,
+                                  <PyObject*>f[0],
+                                  1048576,
+                                  32768,
+                                  131072,
+                                  zran.ZRAN_AUTO_BUILD)
+
+    _zran_init()
+    ret = zran.zran_build_index(&index, 0, 0)
+    assert ret == zran.ZRAN_BUILD_INDEX_OK, ret
+    zran.zran_free(&index)
+
+    _zran_init()
+    ret = zran.zran_seek(&index, dsize - 1, SEEK_SET, NULL)
+    assert ret == zran.ZRAN_SEEK_OK, ret
+    zran.zran_free(&index)
+
+    _zran_init()
+    ret = zran.zran_read(&index, buffer, dsize)
+    assert ret == dsize, ret
+    zran.zran_free(&index)
+
+    pybuf = <bytes>(<char *>buffer)[:dsize]
+    assert np.all(np.frombuffer(pybuf, dtype=np.uint32) == data)
+
+
+# pauldmccarthy/indexed_gzip_fileobj_fork_epicfaace#82
+def test_inflateInit_leak_on_error():
+    """Make sure memory is not leaked after a successful call to
+    inflateInit2(), but then a failure on subsequent zlib calls.
+    """
+
+    cdef zran.zran_index_t index
+
+    # inflateInit2 is called twice in the _zran_zlib_init_inflate function.
+    # We can target the first call by passing a file containing random noise.
+    # I haven't yet figured out a reliable way to target the second call.
+    f = BytesIO(np.arange(1, 100).tobytes())
+
+    iters = np.arange(1, 10000)
+    mem   = np.zeros(10000, dtype=np.uint64)
+
+    for i in iters:
+        assert not zran.zran_init(&index,
+                                  NULL,
+                                  <PyObject*>f,
+                                  1048576,
+                                  32768,
+                                  131072,
+                                  zran.ZRAN_AUTO_BUILD)
+        assert zran.zran_seek(&index, 20, SEEK_SET, NULL) == \
+            zran.ZRAN_SEEK_FAIL
+        zran.zran_free(&index)
+
+        if resource is not None:
+            mem[i] = resource.getrusage(resource.RUSAGE_SELF).ru_maxrss
+
+    # We expect to see some small growth in memory
+    # usage for the first few iterations, but then
+    # it should remain stable
+    mem = mem[5:]
+    assert np.all(mem == mem[0])
+
+
+# pauldmccarthy/indexed_gzip_fileobj_fork_epicfaace#80
+def test_read_eof_memmove_rotate_bug(seed):
+
+    # This bug was triggered by the read buffer rotation
+    # that takes place in zran.c::_zran_read_data_from_file,
+    # and occurs when the file is at EOF, and the
+    # stream->next_in pointer is ahead of index->readbuf by
+    # less than stream->avail_in bytes. In this case, the
+    # source and dest pointers passed to memmove are
+    # overlapping, so the area pointed to by next_in is
+    # modified. The bug was that, when at EOF, the
+    # stream->next_in pointer was not being reset to point
+    # to the beginning of readbuf, so the subsequent read
+    # of the gzip footer in _zran_validate_stream was
+    # reading from the wrong location.
+    #
+    # We can trigger this situation by generating a file
+    # which has compressed file size (X * readbuf_size) + Y,
+    # for any integer x, and for 9 <= Y < 16
+
+    cdef zran.zran_index_t index
+    cdef FILE             *cfid
+
+    with tempdir():
+        nelems = np.random.randint(524288, 525000, 1)[0]
+        data   = np.random.random(nelems)
+        with gzip.open('test.gz', 'wb') as f:
+            f.write(data.tobytes())
+
+        fsize        = os.stat('test.gz').st_size
+        readbuf_size = fsize - 10
+
+        with open('test.gz', 'rb') as pyfid:
+            cfid = fdopen(pyfid.fileno(), 'rb')
+            assert not zran.zran_init(&index,
+                                      cfid,
+                                      NULL,
+                                      4194304,
+                                      32768,
+                                      readbuf_size,
+                                      zran.ZRAN_AUTO_BUILD)
+
+            eof = nelems * 8 - 1
+            got = zran.zran_seek(&index, eof, SEEK_SET, NULL)
+
+            assert got                    == zran.ZRAN_SEEK_OK, got
+            assert zran.zran_tell(&index) == eof